@@ -7,7 +7,7 @@
     "lib/**/!(*.spec).js"
   ],
   "engines": {
-    "node": ">=10"
+    "node": ">=12"
   },
   "scripts": {
     "build": "babel src --out-dir lib --delete-dir-on-start",
@@ -36,18 +36,18 @@
     "enzyme": "^3.6.0",
     "enzyme-adapter-react-16": "^1.5.0",
     "enzyme-adapter-react-helper": "^1.3.0",
-    "eslint": "^7.8.0",
-    "eslint-config-airbnb": "^18.2.0",
-    "eslint-plugin-import": "^2.13.0",
+    "eslint": "^8.4.1",
+    "eslint-config-airbnb": "^19.0.2",
+    "eslint-plugin-import": "^2.25.3",
     "eslint-plugin-jsx-a11y": "^6.1.0",
-    "eslint-plugin-react": "^7.11.0",
+    "eslint-plugin-react": "^7.27.1",
     "jest": "^27.0.3",
     "jest-extended": "^0.11.5",
-    "react": "*",
-    "react-dom": "*",
-    "react-test-renderer": "*"
+    "react": "<17",
+    "react-dom": "<17",
+    "react-test-renderer": "<17"
   },
   "peerDependencies": {
-    "react": "*"
+    "react": "<17"
   }
 }
